splitting functions to mains and aux, fixing __init__ to import all from mains and classes
